+++ src/app/demo/components-overview/components-overview-navigation.service.ts
@@ -96,7 +96,8 @@ export class ComponentsOverviewNavigationService {
       label: 'Badge-Notification',
       moduleName: 'directive'
     },
-    { onclick: () => this.goTo('html'), icon: 'fab fa-html5', label: 'Html', moduleName: 'html' }
+    { onclick: () => this.goTo('html'), icon: 'fab fa-html5', label: 'Html', moduleName: 'html' },
+    { onclick: () => this.goTo('markdown'), icon: 'fab fa-markdown', label: 'Markdown', moduleName: 'markdown' }
   ];
 
   sortedComponentEntries: Map<string, any[]> = new Map<string, any[]>();
@@ -110,6 +111,7 @@ export class ComponentsOverviewNavigationService {
     ['icon', false],
     ['layout', false],
     ['lookup', false],
+    ['markdown', false],
     ['pipes', false],
     ['popup', false]
   ]);